+++ aldrin/src/low_level.rs
@@ -8,6 +8,7 @@ mod event_listener;
 mod promise;
 mod proxy;
 mod reply;
+mod service;
 
 pub(crate) use event_listener::{EventListenerId, EventListenerRequest};
 
@@ -17,3 +18,4 @@ pub use event_listener::EventListener;
 pub use promise::Promise;
 pub use proxy::Proxy;
 pub use reply::Reply;
+pub use service::Service;